requirements.txt
cmake
Werkzeug
Flask
numpy
https://download.pytorch.org/whl/cpu/torch-1.0.1.post2-cp36-cp36m-linux_x86_64.whl
opencv-contrib-python
gevent
Pillow
